Add Fake Analytics instance that uses list to save events sent
